#  Take advantage of JSON.NET PropertyNameTable