--- a/src/Microsoft.AspNetCore.SignalR.Common/Internal/Protocol/JsonHubProtocol.cs
+++ b/src/Microsoft.AspNetCore.SignalR.Common/Internal/Protocol/JsonHubProtocol.cs
@@ -210,7 +210,6 @@ namespace Microsoft.AspNetCore.SignalR.Internal.Protocol
                                     }
                                     else if (FastEquals(reader.Value, HeadersPropertyName))
                                     {
-
                                         JsonUtils.CheckRead(reader);
                                         headers = ReadHeaders(reader);
                                     }