@JamesNK 2018-03-26
Not a fan of reflection on internal types.